附近群组接口tribeId改为long

+++ lbs_web/src/main/java/com/innovate/cms/modules/qs/dao/tribe/ImTribeDao.java
@@ -28,7 +28,7 @@ public interface ImTribeDao extends CrudDao<ImTribe> {
 	 * @param mids
 	 * @return
 	 */
-	List<ImTribe> nearTribe(int[] tribeIds);
+	List<ImTribe> nearTribe(long[] tribeIds);
 
 	/**
 	 * 获取群信息
